1) Add German translation for a string 2) Replace soundcloud logo with a soundcloud lottie animation instead
